[cpp] progress on cpp08, changed unsigned int to size_t pm cpp07/ex02

+++ rank05/cpp/cpp07/ex02/Array.hpp
@@ -8,11 +8,11 @@ template<typename T>
 class Array {
 private:
     T				*array = NULL;
-	unsigned int	arr_size;
+	size_t	arr_size;
 
 public:
     Array() : array(new T()), arr_size(0) {}
-	Array(unsigned int n) : array(new T[n]), arr_size(n) {}
+	Array(size_t n) : array(new T[n]), arr_size(n) {}
 	Array(const Array &other)
 	{
 		*this = other;
@@ -30,7 +30,7 @@ public:
 		// 	delete[] array;
 		T *new_array = new T[other.arr_size];
 
-		for (unsigned int i = 0; i < other.arr_size; i++)
+		for (size_t i = 0; i < other.arr_size; i++)
 			new_array[i] = other.array[i];
 		delete[] array;
 
@@ -39,7 +39,7 @@ public:
 
 		// arr_size = other.arr_size;
 		// array = new T[arr_size];
-		// for (unsigned int i = 0; i < arr_size; i++)
+		// for (size_t i = 0; i < arr_size; i++)
 		// {
 		// 	array[i] = other.array[i]; 
 		// 	// std::cout << "cat" << std::endl;
@@ -47,14 +47,14 @@ public:
 		return (*this);
 	}
 
-    T& operator[](unsigned int index)
+    T& operator[](size_t index)
 	{
 		if (index >= arr_size)
 			throw std::out_of_range("Index is out of range!");
 		return array[index];
     }
 
-    unsigned int size() const
+    size_t size() const
 	{
 		return arr_size;
     }
